+++ lib/reaver/collection.rb
@@ -5,6 +5,7 @@ require 'time'
 require 'digest'
 
 module Reaver
+  # Loading collection from ~/.config/reaver/<filename.yml> and threat them
   class Collection
     attr_reader :tasks
 
@@ -15,39 +16,59 @@ module Reaver
 
     def load_yaml
       puts ">> Loading #{@file}..."
-      if RUBY_VERSION >= '3.0'
-        @tasks = YAML.load_file(@file,  permitted_classes: [Time, Symbol])
-      else
-        @tasks = YAML.load_file(@file)
-      end
-    rescue => error
-      raise error, "loading YAML fail for #{@file}: #{error.message}"
+      @tasks = if RUBY_VERSION >= '3.0'
+                 YAML.load_file(@file, permitted_classes: [Time, Symbol])
+               else
+                 @tasks = YAML.load_file(@file)
+               end
+    rescue => e
+      raise e, "loading YAML fail for #{@file}: #{e.message}"
+    end
+
+    def save_yaml
+      in_yaml = YAML.dump(@tasks)
+      File.write(@file, in_yaml)
     end
 
     def launch(metadata)
-      return unless @tasks
+      return unless @tasks || @tasks['things'].length.zero?
 
-      if @tasks['things'].length >= 1
-        @tasks['things'].each do |t|
-          if File.exist? t['name']
-            old_hash = Digest::MD5.file t['name']
-          else
-            @changed = true
-          end
-
-          Reaver.download(t['url'], t['name']) 
-          compare_hash(t['name'], old_hash) if old_hash
-
-          metadata.info['changed'] = @changed
-        end
+      @tasks['things'].each do |task|
+        do_thing(task)
+        metadata.info['changed'] = @changed
       end
+    end
+
+    protected
+
+    def do_thing(task)
+      hash_exist(task['name'])
+      Reaver.download(task['url'], task['name'])
+      compare_to_old_hash(task['name']) if @old_hash
+      need_to_do_something_with(task) if @changed || !@old_hash
+      @tasks['force_download'] = false
+      @changed = false
     end
 
     private
 
-    def compare_hash(filename, old_hash)
+    def hash_exist(file)
+      if File.exist? file
+        @old_hash = Digest::MD5.file file
+      else
+        @changed = true
+      end
+    end
+
+    def compare_to_old_hash(filename)
       hash = Digest::MD5.file filename
-      @changed = true if old_hash.hexdigest != hash.hexdigest
+      @changed = true if @old_hash.hexdigest != hash.hexdigest
+    end
+
+    def need_to_do_something_with(file)
+      dest = @tasks['all_to_dest_dir'] || file['dest_dir']
+      keep_name = @tasks['keep_name'] || false
+      Reaver::Walk.new(file['name'], dest, keep_name) if dest
     end
   end
 end